+++ src/test/java/task6/MatrixTest.java
@@ -8,8 +8,8 @@ class MatrixTest {
 
     @Test
     public void test() {
-        Matrix m = new Matrix(new int[] {1,2,3,4,5});
-        int[][] expected = new int[][] {
+        Matrix m = new Matrix(new double[] {1,2,3,4,5});
+        double[][] expected = new double[][] {
                 {1,2,3,4,5},
                 {2,3,4,5,1},
                 {3,4,5,1,2},
